+++ api/src/models/User.js
@@ -40,7 +40,6 @@ const userSchema = new Schema({
     },
   ],
   cart: [
-  
     {
       type: Schema.Types.ObjectId,
       ref: "Product",
