+++ api/src/routes/pets/postPet.js
@@ -1,6 +1,6 @@
 const { Router } = require("express");
 const postPet = Router();
-const Pet = require("../models/Pet.js");
+const Pet = require("../../models/Pet.js");
 
 postPet.post("/", async (req, res) => {
   try {
